llmspell_utils

Module system_info

System information gathering utilities System information gathering utilities

This module provides utilities for gathering system information, including OS details, environment variables, and resource usage.

Structs

SystemInfo

System information structure

Enums

OperatingSystem

Operating system type

Functions

find_executable

Find executable in PATH

format_bytes

Format bytes into human-readable string

get_all_env_vars

Get a list of all environment variables

get_cpu_count

Get the number of CPU cores

get_env_or

Get an environment variable with a default value

get_env_var

Get a single environment variable

get_home_directory

Get the user’s home directory

get_hostname

Get the system hostname

get_system_info

Get current system information

get_temp_dir

Get temporary directory path

get_username

Get the current username

is_env_truthy

Check if an environment variable is set and truthy

is_running_in_container

Check if running in a container (Docker, Podman, etc.)

is_running_in_vm

Check if running in a virtual machine

set_env_var_if_allowed

Set an environment variable if allowed